# docs/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/customers/register": {
      "post": {
        "description": "Register a new customer and initiate KYC verification with PartnerCRM. \n    This endpoint works asynchronously - it submits a registration request and returns an acknowledgement \n    with a unique customer GUID. Use the GUID to check registration status using the GET /customers/:guid endpoint.",
        "operationId": "CustomerManagementController_create",
        "parameters": [
          {
            "name": "partner_guid",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bearer_token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "partner",
            "required": true,
            "in": "query",
            "description": "Partner GUID (e.g., \"demo\")",
            "schema": {
              "example": "demo"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerManagementDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer registration request submitted successfully"
          },
          "400": {
            "description": "Invalid request payload - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "error": "ValidationError",
                    "message": "Invalid request data",
                    "details": [
                      {
                        "field": "msisdn",
                        "message": "Phone number must be in international format"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing OAuth token",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "error": "Unauthorized",
                    "message": "Invalid access token"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Customer already exists with this MSISDN or external_identifier",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "error": "ConflictError",
                    "message": "Customer with this MSISDN already exists",
                    "details": {
                      "customer_guid": "cust_1707564600_abc123"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a new customer",
        "tags": [
          "Customer Management"
        ]
      }
    },
    "/api/customers/webhooks/kyc-status": {
      "post": {
        "description": "PartnerCRM sends status updates to this endpoint when customer verification status changes.\n    This endpoint should be configured in PartnerCRM's webhook settings.\n    \n    SECURITY: Validate X-Webhook-Signature header to ensure webhook is from PartnerCRM.",
        "operationId": "CustomerManagementController_receiveKycWebhook",
        "parameters": [
          {
            "name": "x-webhook-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KycWebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "received": true,
                    "message": "Webhook processed successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid webhook payload"
          },
          "401": {
            "description": "Invalid webhook signature"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Webhook callback for KYC status updates",
        "tags": [
          "Customer Management"
        ]
      }
    },
    "/api/payments/initiate": {
      "post": {
        "description": "Initiates payment collection through PaymentHub.\n    Amount must be exactly KES 5,000 for policy premium.",
        "operationId": "PaymentsController_initiatePayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment initiated successfully",
            "schema": {
              "example": {
                "transaction_id": "TXN_123456",
                "status": "pending",
                "amount": 5000,
                "currency": "KES",
                "timestamp": "2026-01-29T10:30:00Z"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payment request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "error": "InvalidAmount",
                    "message": "Payment amount must be exactly KES 5,000 for policy premium",
                    "details": {
                      "expected_amount": 5000,
                      "received_amount": 4000
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "error": "Unauthorized",
                    "message": "Invalid partner authentication token"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Duplicate transaction attempt",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "error": "DuplicateTransaction",
                    "message": "Payment already initiated for this policy",
                    "details": {
                      "existing_transaction_id": "TXN_ASPIN_12345"
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "Payment gateway unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "error": "GatewayTimeout",
                    "message": "PaymentHub gateway temporarily unavailable. Please retry.",
                    "details": {
                      "provider": "mpesa",
                      "timeout_ms": 5000
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initiate payment collection",
        "tags": [
          "Payment Integration"
        ]
      }
    }
  },
  "info": {
    "title": "Aspin Partners and Integrations API Documentation",
    "description": "The Aspin Partners and Integrations API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "swagger",
      "description": "Swagger Documentation"
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "CreateCustomerManagementDto": {
        "type": "object",
        "properties": {
          "msisdn": {
            "type": "string",
            "example": "00254712345678",
            "description": "Customer phone number in international format (MSISDN)"
          },
          "first_name": {
            "type": "string",
            "example": "John",
            "description": "Customer first name"
          },
          "surname": {
            "type": "string",
            "example": "Doe",
            "description": "Customer surname/last name"
          },
          "partner_guid": {
            "type": "string",
            "example": "demo",
            "description": "Partner GUID identifying your organization"
          },
          "display_language": {
            "type": "string",
            "example": "en",
            "description": "Display language code (en, fr, sw, rw, etc.)"
          },
          "national_id": {
            "type": "string",
            "example": "12345678",
            "description": "National ID or passport number"
          },
          "beneficiary_msisdn": {
            "type": "string",
            "example": "00254700000000",
            "description": "Beneficiary phone number in international format"
          },
          "beneficiary_name": {
            "type": "string",
            "example": "Jane Doe",
            "description": "Beneficiary full name"
          },
          "date_of_birth": {
            "type": "string",
            "example": "1990-05-15",
            "description": "Date of birth in YYYY-MM-DD format"
          },
          "external_identifier": {
            "type": "string",
            "example": "ext_cust_12345",
            "description": "External identifier for customer in your system"
          },
          "registration_channel": {
            "type": "string",
            "example": "ApiClient",
            "description": "Registration channel"
          },
          "account_number": {
            "type": "string",
            "example": "1234567890",
            "description": "Bank account number"
          },
          "account_type": {
            "type": "string",
            "example": "Savings",
            "description": "Account type (Savings, Current, etc.)"
          },
          "branch_code": {
            "type": "string",
            "example": "12345",
            "description": "Bank branch code"
          }
        },
        "required": [
          "msisdn",
          "first_name",
          "surname",
          "partner_guid",
          "display_language",
          "national_id",
          "date_of_birth",
          "registration_channel"
        ]
      },
      "VerificationDetailsDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "example": "PartnerCRM"
          },
          "verificationId": {
            "type": "string",
            "example": "kyc_abc123"
          },
          "documentVerified": {
            "type": "boolean",
            "example": true
          },
          "biometricVerified": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "provider",
          "verificationId"
        ]
      },
      "KycWebhookDto": {
        "type": "object",
        "properties": {
          "eventType": {
            "type": "string",
            "enum": [
              "kyc.status.changed"
            ],
            "example": "kyc.status.changed"
          },
          "customer_guid": {
            "type": "string",
            "example": "cust_1234567890_abc",
            "description": "Customer GUID from ASPIN"
          },
          "verificationStatus": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "example": "approved"
          },
          "timestamp": {
            "type": "string",
            "example": "2024-02-10T11:00:00Z"
          },
          "verificationDetails": {
            "$ref": "#/components/schemas/VerificationDetailsDto"
          },
          "rejectionReason": {
            "type": "string",
            "example": "Document quality insufficient"
          }
        },
        "required": [
          "eventType",
          "customer_guid",
          "verificationStatus",
          "timestamp"
        ]
      },
      "InitiatePaymentDto": {
        "type": "object",
        "properties": {
          "policy_code": {
            "type": "string",
            "example": "POL_ASPIN_789012",
            "description": "ASPIN policy identifier requiring premium payment"
          },
          "amount_in_cents": {
            "type": "number",
            "example": 500000,
            "description": "Payment amount in cents (must be exactly 500,000 cents = KES 5,000 for policy premium)",
            "minimum": 500000,
            "maximum": 500000
          },
          "currency": {
            "type": "string",
            "example": "KES",
            "description": "Currency code (ISO 4217)",
            "enum": [
              "KES"
            ]
          },
          "provider": {
            "type": "string",
            "example": "mpesa",
            "description": "Payment gateway provider",
            "enum": [
              "mpesa",
              "airtel"
            ]
          },
          "msisdn": {
            "type": "string",
            "example": "254712345678",
            "description": "Customer MSISDN in E.164 format (without + or 00 prefix)",
            "pattern": "^254[0-9]{9}$"
          },
          "channel": {
            "type": "string",
            "example": "APIClient",
            "description": "Payment channel identifier",
            "enum": [
              "APIClient",
              "WebPortal",
              "MobileApp",
              "AgentPortal"
            ]
          },
          "product_code": {
            "type": "string",
            "example": "PROD_HEALTH_001",
            "description": "Product code identifying the insurance product"
          },
          "aspin_reference": {
            "type": "string",
            "example": "ASP_REF_20260211_12345",
            "description": "ASPIN-generated reference for idempotency (recommended)"
          },
          "description": {
            "type": "string",
            "example": "Premium payment for Policy POL_ASPIN_789012",
            "description": "Payment description shown to customer"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-02-11T14:30:00Z",
            "description": "Timestamp of payment request (ISO 8601)"
          }
        },
        "required": [
          "policy_code",
          "amount_in_cents",
          "currency",
          "provider",
          "msisdn",
          "channel",
          "product_code"
        ]
      },
      "PaymentResponseDto": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string",
            "example": "TXN_ASPIN_20260211_7890",
            "description": "ASPIN-generated transaction reference (use for status checks)"
          },
          "status": {
            "type": "string",
            "example": "pending",
            "description": "Current transaction status",
            "enum": [
              "pending",
              "completed",
              "failed"
            ]
          },
          "amount": {
            "type": "number",
            "example": 5000,
            "description": "Payment amount in KES"
          },
          "currency": {
            "type": "string",
            "example": "KES",
            "description": "Currency code"
          },
          "provider": {
            "type": "string",
            "example": "mpesa",
            "description": "Payment gateway used",
            "enum": [
              "mpesa",
              "airtel"
            ]
          },
          "timestamp": {
            "type": "string",
            "example": "2026-02-11T14:30:00Z",
            "description": "Transaction initiation timestamp (ISO 8601)"
          },
          "policy_id": {
            "type": "string",
            "example": "POL_ASPIN_789012",
            "description": "Associated ASPIN policy ID"
          },
          "aspin_reference": {
            "type": "string",
            "example": "ASP_REF_20260211_12345",
            "description": "ASPIN reference for idempotency"
          },
          "message": {
            "type": "string",
            "example": "Payment initiated successfully. Customer will receive prompt.",
            "description": "Human-readable status message"
          }
        },
        "required": [
          "transaction_id",
          "status",
          "amount",
          "currency",
          "provider",
          "timestamp",
          "policy_id",
          "message"
        ]
      }
    }
  }
}